requirements.txt
torch
diffusers
transformers
safetensors
Pillow
huggingface_hub
python-dotenv
